perubahan tempat dan rename class
