+++ local_art_and_culture/lib/src/feature/festival/screen/invoice.dart
@@ -100,7 +100,7 @@ class _InvoiceScreenState extends State<InvoiceScreen> {
                     Navigator.pushAndRemoveUntil(
                         context,
                         MaterialPageRoute(
-                            builder: (context) => const HomeScreen()),
+                            builder: (context) => const HomeScreenFstv()),
                         (route) => false);
                   },
                   style: ElevatedButton.styleFrom(
